+++ AppVersion.cs
@@ -2,8 +2,8 @@ namespace NOCS
 {
     internal static class AppVersion
     {
-        public const string ReleaseBase = "0.5.26";
+        public const string ReleaseBase = "0.5.27";
         public const string BepInSemVer = ReleaseBase;
-        public const string DisplayVersion = "0.5.26QV";
+        public const string DisplayVersion = "0.5.27QV";
     }
 }